requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==5.9.0
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-js-asset==1.2.2
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
Unidecode==1.1.1
